Moved HeaderBar.jsx and StatusIndicator.jsx into general directory
