+++ frontend/src/streamlist/SingleStreamRow.jsx
@@ -10,7 +10,7 @@ import {
 import { ExpandLess, ExpandMore } from "@material-ui/icons/";
 
 import PropTypes from "prop-types";
-import StatusIndicator from "../StatusIndicator";
+import StatusIndicator from "../general/StatusIndicator";
 import ActionButtons from "./ActionButtons";
 import StreamInfo from "../model/StreamInfo";
